refactor playground status chip component
