+++ frontend/src/component/playground/Playground/PlaygroundResultsTable/PlaygroundResultChip/PlaygroundResultChip.tsx
@@ -1,10 +1,10 @@
 import { VFC } from 'react';
-import { Chip, styled, useTheme } from '@mui/material';
-import { colors } from 'themes/colors';
+import { useTheme } from '@mui/material';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
 import { ReactComponent as FeatureEnabledIcon } from 'assets/icons/isenabled-true.svg';
 import { ReactComponent as FeatureDisabledIcon } from 'assets/icons/isenabled-false.svg';
 import { WarningOutlined } from '@mui/icons-material';
+import { StatusChip } from 'component/common/StatusChip/StatusChip';
 
 interface IResultChipProps {
     enabled: boolean | 'unevaluated' | 'unknown';
@@ -12,51 +12,6 @@ interface IResultChipProps {
     // Result icon - defaults to true
     showIcon?: boolean;
 }
-
-export const StyledChip = styled(Chip)(({ theme, icon }) => ({
-    padding: theme.spacing(0, 1),
-    height: 24,
-    borderRadius: theme.shape.borderRadius,
-    fontWeight: theme.typography.fontWeightMedium,
-    ['& .MuiChip-label']: {
-        padding: 0,
-        paddingLeft: Boolean(icon) ? theme.spacing(0.5) : 0,
-    },
-}));
-
-export const StyledFalseChip = styled(StyledChip)(({ theme }) => ({
-    border: `1px solid ${theme.palette.error.main}`,
-    backgroundColor: colors.red['200'],
-    ['& .MuiChip-label']: {
-        color: theme.palette.error.main,
-    },
-    ['& .MuiChip-icon']: {
-        color: theme.palette.error.main,
-    },
-}));
-
-export const StyledTrueChip = styled(StyledChip)(({ theme }) => ({
-    border: `1px solid ${theme.palette.success.main}`,
-    backgroundColor: colors.green['100'],
-    ['& .MuiChip-label']: {
-        color: theme.palette.success.main,
-    },
-    ['& .MuiChip-icon']: {
-        color: theme.palette.success.main,
-        marginRight: 0,
-    },
-}));
-
-export const StyledUnknownChip = styled(StyledChip)(({ theme }) => ({
-    border: `1px solid ${theme.palette.warning.main}`,
-    backgroundColor: colors.orange['100'],
-    ['& .MuiChip-label']: {
-        color: theme.palette.warning.main,
-    },
-    ['& .MuiChip-icon']: {
-        color: theme.palette.warning.main,
-    },
-}));
 
 export const PlaygroundResultChip: VFC<IResultChipProps> = ({
     enabled,
@@ -92,22 +47,21 @@ export const PlaygroundResultChip: VFC<IResultChipProps> = ({
         <ConditionallyRender
             condition={enabled === 'unknown' || enabled === 'unevaluated'}
             show={
-                <StyledUnknownChip
-                    icon={showIcon ? icon : undefined}
-                    label={label}
-                />
+                <StatusChip icon={showIcon ? icon : undefined} label={label} />
             }
             elseShow={
                 <ConditionallyRender
                     condition={typeof enabled === 'boolean' && Boolean(enabled)}
                     show={
-                        <StyledTrueChip
+                        <StatusChip
+                            variant="true"
                             icon={showIcon ? icon : undefined}
                             label={label}
                         />
                     }
                     elseShow={
-                        <StyledFalseChip
+                        <StatusChip
+                            variant="false"
                             icon={showIcon ? icon : undefined}
                             label={label}
                         />
